chore: move sidebar to its component
