+++ app/containers/App/index.tsx
@@ -6,61 +6,25 @@
  * contain code that should be seen on all pages. (e.g. navigation bar)
  *
  */
-import React, { useState, useCallback, useEffect } from 'react';
+import React, { useState, useCallback } from 'react';
 import GlobalStyle from '@app/global-styles';
 import { routeConfig } from '@app/routeConfig';
-import { Layout, Drawer, DrawerProps } from 'antd';
+import { Layout } from 'antd';
 import map from 'lodash-es/map';
 import { withRouter } from 'react-router';
-import { Link, Route, Switch } from 'react-router-dom';
+import { Route, Switch } from 'react-router-dom';
 import { compose } from 'redux';
 import styled, { ThemeProvider } from 'styled-components';
 import For from '@components/For';
 import Header from '@components/Header';
 import { colors } from '@themes/index';
-import icon from '@images/ion_rocket-sharp.svg';
-import { CloseOutlined } from '@ant-design/icons';
+import Sidebar from '@app/components/Siderbar';
+import useMobile from '@utils/useMobile';
 
 const theme = {
   fg: colors.primary,
   bg: colors.secondaryText
 };
-
-const SidebarDrawer = styled(Drawer)`
-  && {
-    .ant-drawer-body {
-      padding: 7rem 0 0 0;
-      background-color: ${colors.primary};
-      width: 6%;
-      min-width: 4.5rem;
-      max-width: 7rem;
-      text-align: center;
-    }
-    .ant-drawer-close {
-      top: 1rem;
-    }
-  }
-`;
-
-const SideBar = styled.div`
-  && {
-    width: 6%;
-    min-width: 4.5rem;
-    max-width: 7rem;
-    min-height: calc(100vh - 7rem);
-    height: auto;
-    background-color: ${colors.primary};
-    display: inline;
-    text-align: center;
-  }
-`;
-
-const RocketLogo = styled.img`
-  && {
-    margin-top: 1rem;
-    object-fit: contain;
-  }
-`;
 
 const CustomLayout = styled(Layout)`
   && {
@@ -72,48 +36,15 @@ export const MOBILE_BREAKPOINT = 450;
 
 export function App() {
   const [visible, setVisible] = useState(false);
-  const [mobile, setMobile] = useState(window.innerWidth < MOBILE_BREAKPOINT);
-
-  function detectMobile() {
-    if (window.innerWidth < MOBILE_BREAKPOINT) {
-      setMobile(true);
-    } else {
-      setMobile(false);
-    }
-  }
-
-  useEffect(() => {
-    window.addEventListener('resize', detectMobile);
-    return () => {
-      window.removeEventListener('resize', detectMobile);
-    };
-  }, []);
+  const { mobile } = useMobile(MOBILE_BREAKPOINT);
 
   const toggleSidebar = useCallback(() => setVisible((v) => !v), []);
-
-  const sidebarProps: DrawerProps = mobile
-    ? {
-        closeIcon: <CloseOutlined style={{ color: colors.secondary, fontSize: '1.9rem' }} />,
-        placement: 'left',
-        visible,
-        closable: true,
-        onClose: toggleSidebar,
-        width: 'max-content'
-      }
-    : {};
-
-  const SidebarComponent = (props: DrawerProps) =>
-    mobile ? <SidebarDrawer {...props} /> : <SideBar data-testid="sidebar" {...(props as any)} />;
 
   return (
     <ThemeProvider theme={theme}>
       <Header mobile={mobile} toggleSidebar={toggleSidebar} />
       <CustomLayout>
-        <SidebarComponent {...sidebarProps}>
-          <Link onClick={toggleSidebar} data-testid="rocket-home-link" aria-label="home link" to="/">
-            <RocketLogo src={icon} alt="rocket-icon" />
-          </Link>
-        </SidebarComponent>
+        <Sidebar toggleSidebar={toggleSidebar} mobile={mobile} visible={visible} />
         <Layout.Content>
           <For
             ParentComponent={(props) => <Switch {...props} />}
